Fix the keyboard navigation which was broken in flipped orientations
When orientation is HorizontalFlipped, reverse the meaning of a and d
keys.  When orientation is VerticalFlipped, reverse the meaning of w
and s keys.

I decided that  making the change in the key handling code in
insertText:(NSString*)theText was more intuitive than actually changing
the meaning of left, right, up, and down in the move* methods.

My original intent with the flipped layout code was to make the change
as noninvasive as possible, this change is in keeping with that goal.